Set ext_client mtu to match ingress gw
Currently the external client configuration does not specify a MTU,
which causes WireGuard to use the default of 1420. The default of the
ingress gateway will be 1280. This mismatch can cause network issues.

This patch ensures that the generated external client configuration is
specified to use a MTU that matches that one the ingress gateway.