all: EdDSA replacement for password authentication
    This is the initial scaffolding for replacing password-based
    authentication with digital signature based authentication.

This patch uses  EdDSA as the chosen signature scheme, however this is not set in stone.

The rationale behind selecting EdDSA is based on the compatibility with existing WireGuard keys. WireGuard uses a 1-RTT key exchange, with Curve25519 [1] for ECDH. The curve25519[2] coordinates used in ECDH are in Montgomery form, EdDSA uses coordinates in the corresponding Edwards representation.

There exists a birational equivalence[2] between Edwards and Montgomery curves. Relevant to us, and in short, this means existing WireGuard keys (curve25519 points in their Montgomery form used for ECDH) are computationally transformed to "EdDSA keys" (curve25519 points in their Twisted Edwards form used for EdDSA).

While EdwardsToMontgomery is an available transformation, (and what this patch uses) MontgomeryToEdwards is not implemented here and is a future TODO.

One concern about going from Montgomery to Edwards is that unlike the inverse transformation, two unique coordinate pairs exist in Edwards form for
each coordinate pair in Montgomery form. Which point to select is
seemingly non-standard as reflected in the differences in point
selection between, for example, NaCL and other libraries.

[1]: https://www.wireguard.com/papers/wireguard.pdf
[2]: https://cr.yp.to/ecdh/curve25519-20060209.pdf
[3]: https://en.wikipedia.org/wiki/Montgomery_curve#Equivalence_with_twisted_Edwards_curves

* workflows/test: bump go1.17 -> go1.18

Signed-off-by: John Sahhar <john@gravitl.com>